+++ cmd/test/main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"log"
+	"log/slog"
 	"os"
 	"os/signal"
 	"syscall"
@@ -12,15 +13,31 @@ import (
 	commonVaultTokenClient "github.com/crypto-bundle/bc-wallet-common-lib-vault/pkg/vault/client/token"
 )
 
+var ErrMockFormatter = errors.New("mock_err_formatter")
+
+type loggBuilder struct{}
+
+func (f *loggBuilder) NewSlogLoggerEntry(fields ...any) *slog.Logger {
+	return slog.Default()
+}
+
+func (f *loggBuilder) NewSlogNamedLoggerEntry(named string, fields ...any) *slog.Logger {
+	return slog.Default()
+}
+
+func (f *loggBuilder) NewSlogLoggerEntryWithFields(fields ...slog.Attr) *slog.Logger {
+	return slog.Default()
+}
+
 type errFmt struct {
 }
 
 func (f *errFmt) ErrorWithCode(_ error, _ int) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) ErrWithCode(_ error, _ int) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) ErrorGetCode(_ error) int {
@@ -32,34 +49,40 @@ func (f *errFmt) ErrGetCode(_ error) int {
 }
 
 func (f *errFmt) ErrorNoWrap(_ error) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) ErrNoWrap(_ error) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) ErrorOnly(_ error, _ ...string) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) Error(_ error, _ ...string) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) Errorf(_ error, _ string, _ ...interface{}) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) NewError(_ ...string) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func (f *errFmt) NewErrorf(_ string, _ ...interface{}) error {
-	return errors.New("mock_err_formatter")
+	return ErrMockFormatter
 }
 
 func main() {
+	const (
+		DefaultTokenRenewTTL        = 960
+		DefaultVaultConnectionPOrt  = 8200
+		DefaultOsSignalsChannelSize = 2
+	)
+
 	type VaultWrappedConfig struct {
 		*commonVault.BaseConfig
 		*commonVaultTokenClient.AuthConfig
@@ -68,10 +91,10 @@ func main() {
 	vaultCfg := &VaultWrappedConfig{
 		BaseConfig: &commonVault.BaseConfig{
 			Host:          "127.0.0.1",
-			Port:          8200,
+			Port:          DefaultVaultConnectionPOrt,
 			UseHTTPS:      false,
 			AuthMethod:    "token",
-			TokenRenewTTL: 960,
+			TokenRenewTTL: DefaultTokenRenewTTL,
 			DataPath:      "kv/data/crypto-bundle/bc-wallet-common/transit,kv/data/crypto-bundle/bc-wallet-ethereum-hdwallet/common",
 		},
 		AuthConfig: &commonVaultTokenClient.AuthConfig{
@@ -86,6 +109,7 @@ func main() {
 	}
 
 	mockErrFmtSvc := &errFmt{}
+	mockLoggBuilder := &loggBuilder{}
 	ctx, cancelCtxFunc := context.WithCancel(context.Background())
 
 	vaultClientSrv, err := commonVaultTokenClient.NewClient(ctx, mockErrFmtSvc, vaultCfg)
@@ -93,7 +117,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	vaultSrv, err := commonVault.NewService(log.Default(), mockErrFmtSvc, vaultCfg, vaultClientSrv)
+	vaultSrv, err := commonVault.NewService(mockLoggBuilder, mockErrFmtSvc, vaultCfg, vaultClientSrv)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -103,12 +127,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	//err = vaultSrv.LoadSecrets(ctx)
-	//if err != nil {
-	//	log.Fatal(err)
-	//}
-
-	c := make(chan os.Signal, 2)
+	c := make(chan os.Signal, DefaultOsSignalsChannelSize)
 	signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
 	<-c
 	cancelCtxFunc()