+++ pkg/vault/client/kubernates/client.go
@@ -8,15 +8,25 @@ import (
 	k8sAuth "github.com/hashicorp/vault/api/auth/kubernetes"
 )
 
+const AuthMethodName = "kubernetes"
+
 var (
+	_ selfService = (*service)(nil)
+
 	ErrEmptySecret = errors.New("unable to get secret")
 )
 
 type service struct {
+	e   errorFormatterService
+	cfg configService
+
 	vaultConfig *vaultApi.Config
 	client      *vaultApi.Client
 	k8sAuth     *k8sAuth.KubernetesAuth
-	cfg         configService
+}
+
+func (s *service) GetAuthMethod() string {
+	return AuthMethodName
 }
 
 func (s *service) GetClient() *vaultApi.Client {
@@ -26,10 +36,11 @@ func (s *service) GetClient() *vaultApi.Client {
 func (s *service) Login(ctx context.Context) (*vaultApi.Client, error) {
 	authInfo, err := s.client.Auth().Login(ctx, s.k8sAuth)
 	if err != nil {
-		return nil, err
+		return nil, s.e.ErrorOnly(err)
 	}
+
 	if authInfo == nil {
-		return nil, ErrEmptySecret
+		return nil, s.e.ErrorOnly(ErrEmptySecret)
 	}
 
 	s.client.SetToken(authInfo.Auth.ClientToken)
@@ -38,13 +49,16 @@ func (s *service) Login(ctx context.Context) (*vaultApi.Client, error) {
 }
 
 // NewClient initialize vault client with service account token authorization.
-func NewClient(_ context.Context, cfg configService) (*service, error) {
+func NewClient(_ context.Context,
+	errFmtSvc errorFormatterService,
+	cfg configService,
+) (*service, error) {
 	clientOpts := vaultApi.DefaultConfig()
 	clientOpts.Address = cfg.GetAddress()
 
 	client, err := vaultApi.NewClient(clientOpts)
 	if err != nil {
-		return nil, err
+		return nil, errFmtSvc.ErrorOnly(err)
 	}
 
 	auth, err := k8sAuth.NewKubernetesAuth(
@@ -53,13 +67,16 @@ func NewClient(_ context.Context, cfg configService) (*service, error) {
 		k8sAuth.WithMountPath(cfg.GetKubernatesAuthPath()),
 	)
 	if err != nil {
-		return nil, err
+		return nil, errFmtSvc.ErrorOnly(err)
 	}
 
 	vaultSvc := &service{
-		client:  client,
-		cfg:     cfg,
-		k8sAuth: auth,
+		e:   errFmtSvc,
+		cfg: cfg,
+
+		client:      client,
+		k8sAuth:     auth,
+		vaultConfig: nil,
 	}
 
 	return vaultSvc, nil