+++ pkg/vault/client/userpass/client.go
@@ -8,16 +8,25 @@ import (
 	userpassAuth "github.com/hashicorp/vault/api/auth/userpass"
 )
 
+const AuthMethodName = "userpass"
+
 var (
+	_ selfService = (*service)(nil)
+
 	ErrEmptySecret = errors.New("unable to get secret")
 )
 
 type service struct {
+	e   errorFormatterService
 	cfg configService
 
 	auth        *userpassAuth.UserpassAuth
 	vaultConfig *vaultApi.Config
 	client      *vaultApi.Client
+}
+
+func (s *service) GetAuthMethod() string {
+	return AuthMethodName
 }
 
 func (s *service) GetClient() *vaultApi.Client {
@@ -27,10 +36,11 @@ func (s *service) GetClient() *vaultApi.Client {
 func (s *service) Login(ctx context.Context) (*vaultApi.Client, error) {
 	secret, err := s.client.Auth().Login(ctx, s.auth)
 	if err != nil {
-		return nil, err
+		return nil, s.e.ErrorOnly(err)
 	}
+
 	if secret == nil {
-		return nil, ErrEmptySecret
+		return nil, s.e.ErrorOnly(ErrEmptySecret)
 	}
 
 	s.client.SetToken(secret.Auth.ClientToken)
@@ -40,22 +50,32 @@ func (s *service) Login(ctx context.Context) (*vaultApi.Client, error) {
 
 // NewClient initialize vault client with user and password
 // authentication.
-func NewClient(ctx context.Context, cfg configService) (*service, error) {
+func NewClient(_ context.Context,
+	errFmtSvc errorFormatterService,
+	cfg configService,
+) (*service, error) {
 	clientOpts := vaultApi.DefaultConfig()
 	clientOpts.Address = cfg.GetAddress()
+
 	client, err := vaultApi.NewClient(clientOpts)
 	if err != nil {
-		return nil, err
+		return nil, errFmtSvc.ErrorOnly(err)
 	}
 
 	auth, err := userpassAuth.NewUserpassAuth(cfg.GetUserName(), &userpassAuth.Password{
+		FromFile:   "",
+		FromEnv:    "",
 		FromString: cfg.GetUserPassword(),
 	})
 	if err != nil {
-		return nil, err
+		return nil, errFmtSvc.ErrorOnly(err)
 	}
 
 	return &service{
+		e: errFmtSvc,
+
+		vaultConfig: nil,
+
 		client: client,
 		cfg:    cfg,
 		auth:   auth,