+++ pkg/vault/common.go
@@ -2,6 +2,7 @@ package vault
 
 import (
 	"context"
+	"log/slog"
 
 	vaultApi "github.com/hashicorp/vault/api"
 )
@@ -18,34 +19,35 @@ type configService interface {
 	// Token auth config methods
 	// --------------------------------------------
 
-	//GetAuthToken() string
+	// GetAuthToken() string
 	// --------------------------------------------
 	// Github token auth config methods
 	// --------------------------------------------
 
-	//GetGithubAuthPath() string
-	//GetGithubAuthToken() string
+	// GetGithubAuthPath() string
+	// GetGithubAuthToken() string
 	// --------------------------------------------
 	// User and password auth config methods
 
-	//GetUserName() string
-	//GetUserPassword() string
+	// GetUserName() string
+	// GetUserPassword() string
 
 	// --------------------------------------------
 	// Kubernates config methods
 	// --------------------------------------------
 
-	//GetKubernatesAppRole() string
-	//GetKubernatesSATokenPath() string
-	//GetKubernatesAuthPath() string
+	// GetKubernatesAppRole() string
+	// GetKubernatesSATokenPath() string
+	// GetKubernatesAuthPath() string
 
 	GetApplicationStageName() string
 	GetApplicationEnvironment() string
 }
 
 type clientService interface {
+	GetAuthMethod() string
 	GetClient() *vaultApi.Client
-	Login(context.Context) (*vaultApi.Client, error)
+	Login(ctx context.Context) (*vaultApi.Client, error)
 }
 
 type renewerService interface {
@@ -58,14 +60,22 @@ type baseApplicationConfigService interface {
 	GetStageName() string
 }
 
-type Vaulter interface {
+type Encryptor interface {
 	Encrypt(toEncrypt []byte) ([]byte, error)
 	Decrypt(cipherBytes []byte) ([]byte, error)
+}
 
+type Vaulter interface {
 	GetCredentialsBytes() (b []byte, err error)
 	GetCredentialsBytesByPath(path string) (b []byte, err error)
 	GetCredentialsByPathAndKey(path, field string) (string, error)
 	GetCredentialsByPathAndKeys(path string, fields ...string) (map[string]string, error)
+}
+
+type loggerFabricService interface {
+	NewSlogLoggerEntry(fields ...any) *slog.Logger
+	NewSlogNamedLoggerEntry(named string, fields ...any) *slog.Logger
+	NewSlogLoggerEntryWithFields(fields ...slog.Attr) *slog.Logger
 }
 
 type errorFormatterService interface {