+++ pkg/vault/renewer.go
@@ -2,7 +2,7 @@ package vault
 
 import (
 	"context"
-	"log"
+	"log/slog"
 	"sync"
 
 	vaultApi "github.com/hashicorp/vault/api"
@@ -17,13 +17,13 @@ const (
 )
 
 type renewer struct {
-	l *log.Logger
+	l *slog.Logger
 	e errorFormatterService
 
-	client     clientService
-	defaultTTL int
-
+	client        clientService
 	currentSecret *vaultApi.Secret
+
+	defaultTTL int
 }
 
 func (s *renewer) IsHealed(_ context.Context) bool {
@@ -53,7 +53,7 @@ func (s *renewer) prepareRenew(_ context.Context) error {
 
 	secret, loopErr := token.RenewSelf(s.defaultTTL)
 	if loopErr != nil {
-		s.l.Printf("vault token renew error: %v", loopErr)
+		s.l.Error("vault token renew error", loopErr)
 
 		return s.e.ErrorOnly(loopErr, ErrUnableGetTokenInfoDetail)
 	}
@@ -71,45 +71,48 @@ func (s *renewer) PrepareAndStartRenew(ctx context.Context) error {
 
 	wg := &sync.WaitGroup{}
 	wg.Add(1)
+
 	go func() {
-		err = s.startRenew(ctx, sync.OnceFunc(func() {
+		s.startRenew(ctx, sync.OnceFunc(func() {
 			wg.Done()
 		}))
 	}()
-	wg.Wait()
 
-	if err != nil {
-		return s.e.ErrorNoWrap(err)
-	}
+	wg.Wait()
 
 	return nil
 }
 
-func (s *renewer) startRenew(ctx context.Context, renewClb func()) error {
+func (s *renewer) startRenew(ctx context.Context, renewClb func()) {
 	for {
 		renewed, loopErr := s.renew(ctx, renewClb)
 		if loopErr != nil {
-			s.l.Printf("vault token renew error: %v", loopErr)
+			s.l.Error("vault token renew error", loopErr)
 		}
 
 		if renewed&exitRequested != 0 {
-			return nil
+			return
 		}
 
 		if renewed&expiringAuthToken != 0 {
 			_, loginErr := s.client.Login(ctx)
 			if loginErr != nil {
-				s.l.Printf("login authentication error: %v", loginErr)
+				s.l.Error("login authentication error", loginErr)
 			}
 
-			s.l.Printf("reconnect and renew")
+			s.l.Info("reconnect and renew")
 		}
 	}
 }
 
 func (s *renewer) renew(ctx context.Context, renewClb func()) (renewResult, error) {
 	authTokenWatcher, err := s.client.GetClient().NewLifetimeWatcher(&vaultApi.LifetimeWatcherInput{
-		Secret: s.currentSecret,
+		Secret:        s.currentSecret,
+		Grace:         0,
+		Rand:          nil,
+		RenewBuffer:   0,
+		Increment:     0,
+		RenewBehavior: 0,
 	})
 	if err != nil {
 		return renewError, s.e.ErrorOnly(err, ErrInitTokenTTLWatcherDetail)
@@ -126,21 +129,21 @@ func (s *renewer) renew(ctx context.Context, renewClb func()) (renewResult, erro
 			return exitRequested, nil
 
 		case doneErr, isClosed := <-authTokenWatcher.DoneCh():
-			s.l.Printf("auth token: done with err: %s and chan is closed: %t",
-				doneErr, isClosed)
+			s.l.Error("renew auth token done with err, and chan is closed",
+				doneErr, slog.Bool(RenewTokenChannelStatusTag, isClosed))
 
 			return expiringAuthToken, s.e.ErrorNoWrap(doneErr)
 
 		case info := <-authTokenWatcher.RenewCh():
 			s.currentSecret = info.Secret
 
-			s.l.Printf("auth token: successfully renewed; remaining duration: %ds",
-				info.Secret.Auth.LeaseDuration)
+			s.l.Info("successfully renewed",
+				slog.Int(RenewTokenLeaseDurationTag, info.Secret.Auth.LeaseDuration))
 		}
 	}
 }
 
-func newRenewer(logger *log.Logger,
+func newRenewer(logger *slog.Logger,
 	errFmtSvc errorFormatterService,
 	clientSvc clientService,
 	defaultTTL int,